+++ providers/RequestExtensionProvider.js
@@ -1,9 +1,27 @@
 'use strict'
 
+const murmurhash = require('murmurhash-native')
 const { ServiceProvider } = require('@adonisjs/fold')
 const url = require('url')
 
 class RequestExtensionProvider extends ServiceProvider {
+/**
+ * Setup 'Cache-Control' response headers
+ * taking into consideration the request
+ * 
+ * @param {Object} ctx 
+ * @param {Function} next 
+ * @param {Array} [-destructured-]
+ * 
+ * @method cacheHeadersMiddleware
+ * 
+ * @return {Promise}
+ */
+ async cacheHeadersMiddleware(ctx, next, [directives = 'no-store,max-age=0']){
+
+    let cacheInstructions = ctx.request.header('Cache-Control')
+ }
+
 /**
  * Register namespaces to the IoC container
  *
@@ -25,6 +43,7 @@ class RequestExtensionProvider extends ServiceProvider {
  */
   boot () {
     const Request = this.app.use('Adonis/Src/Request')
+    const Server = this.app.use('Server')
     const RouteManager = this.app.use('Adonis/Src/Route')
 
     Request.getter('currentTime', function () {
@@ -43,7 +62,27 @@ class RequestExtensionProvider extends ServiceProvider {
       let urlParts = url.parse(this.url(), true)
       let route = RouteManager.match(urlParts.pathname, this.method().toUpperCase(), urlParts.host)
 
-      return route !== null ? route.route.toJSON() : { name: '', route: null, verbs: [], middleware: [], handler: null, domain: null }
+      let current = ( route !== null ? route.route.toJSON() : { name: '', route: null, verbs: [], middleware: [], handler: null, domain: null })
+
+      current.isNamed = function(...patterns){
+
+      };
+
+      return current
+    })
+
+    Request.macro('fingerprint', function (unique = true) {
+       // If requests exceed 1 billion, collion make occur (rate:2%)
+      let currentRoute = this.currentRoute()
+
+      if(currentRoute.route === null){
+         return null;
+      }
+
+      return murmurhash.murmurHash64x86([].concat(
+        (unique === true ? [this.method()] : currentRoute.verbs),
+        [currentRoute.domain, this.url(), this.ip()]
+      ).join('|'));
     })
 
     Request.macro('port', function () {
@@ -69,6 +108,8 @@ class RequestExtensionProvider extends ServiceProvider {
     Request.macro('hasHeader', function (headerText) {
       return (typeof this.header(headerText) === 'string')
     })
+
+    Server.registerNamed({ 'cache.headers': this.cacheHeadersMiddleware })
   }
 }