+++ CHANGELOG.md
@@ -1,8 +1,21 @@
+<a name="0.0.6"></a>
+# 0.0.6 (2019-05-30)
+
+### Bug Fixes
+- Fixed 
+
+### Features Added
+- Added `request.fingerprint()`
+- Added `request.currentRoute().isNamed()`
+
+
 <a name="0.0.5"></a>
 # 0.0.5 (2019-04-14)
 
 ### Bug Fixes
 - Fixed logic error `toButton()`
+
+### Features Added
 - Added `request.origin()`
 
 <a name="0.0.4"></a>
@@ -10,6 +23,8 @@
 
 ### Bug Fixes
 - Fixed UpdateViewData Middleware Syntax error
+
+### Features Added
 - Added View global `toButton()`
 
 <a name="0.0.3"></a>